refactor: remove unused code and move shared components to common package
