+++ webtrit_callkeep_android/android/src/main/kotlin/com/webtrit/callkeep/services/common/IsolateSelectionStrategy.kt
@@ -1,4 +1,4 @@
-package com.webtrit.callkeep.services.services.incoming_call
+package com.webtrit.callkeep.services.common
 
 import androidx.lifecycle.Lifecycle
 import com.webtrit.callkeep.common.Log
